requirements.txt
fastapi
uvicorn
python-multipart
aiofiles
pydantic
redis
celery
google-genai
requests
dotenv
sqlalchemy 
asyncpg 
argon2-cffi
python-jose
